+++ packages/spotlight/CHANGELOG.md
@@ -15,6 +15,8 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ### Fixed
 
+- `spotlight` to correctly restore focus to a spotlight container in another container
+
 ## [1.6.1] - 2017-08-07
 
 No significant changes.