+++ packages/spotlight/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 The following is a curated list of changes in the Enact spotlight module, newest changes on the top.
 
+## [1.7.0] - 2017-08-23
+
+### Added
+
+- `spotlight/SpotlightContainerDecorator` config option `continue5WayHold` to support moving focus to the next spottable element on 5-way hold key.
+- `spotlight/Spottable` ability to restore focus when an initially disabled component becomes enabled
+
+### Fixed
+
+- `spotlight` to correctly restore focus to a spotlight container in another container
+- `spotlight` to not try to focus something when the window is activated if focus is already set
+
+## [1.6.1] - 2017-08-07
+
+No significant changes.
+
 ## [1.6.0] - 2017-08-04
 
 ### Changed